LayoutTests/fast/canvas/webgl/getBufferSubData-webgl1-expected.txt

CONSOLE MESSAGE: line 40: TypeError: null is not an object (evaluating 'context.createBuffer')
Make sure that getBufferSubData() doesn't work on WebGL 1 contexts

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
